move utils to crypto
