+++ examples/import-in-node/.env.local.example
@@ -3,4 +3,3 @@ API_PRIVATE_KEY="<Turnkey API Private Key>"
 USER_ID="<Turnkey User ID>"
 BASE_URL="https://api.turnkey.com"
 ORGANIZATION_ID="<Turnkey organization ID>"
-SIGNER_PUBLIC_KEY="<04f3422b8afbe425d6ece77b8d2469954715a2ff273ab7ac89f1ed70e0a9325eaa1698b4351fd1b23734e65c0b6a86b62dd49d70b37c94606aac402cbd84353212 for preprod, 04cf288fe433cc4e1aa0ce1632feac4ea26bf2f5a09dcfe5a42c398e06898710330f0572882f4dbdf0f5304b8fc8703acd69adca9a4bbf7f5d00d20a5e364b2569 for prod>"
