+++ examples/import-in-node/package.json
@@ -18,7 +18,6 @@
     "@turnkey/encoding": "workspace:*",
     "@turnkey/sdk-server": "workspace:*",
     "async-retry": "^1.3.3",
-    "bs58": "^5.0.0",
     "cross-fetch": "^4.0.0",
     "dotenv": "^16.0.3"
   },
